+++ src/main/java/com/woowa/rentcar/view/InputView.java
@@ -1,0 +1,4 @@
+package com.woowa.rentcar.view;
+
+public class InputView {
+}
