+++ src/test/java/com/woowa/rentcar/AvanteTest.java
@@ -1,9 +1,10 @@
 package com.woowa.rentcar;
 
+import com.woowa.rentcar.domain.Avante;
+
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 class AvanteTest {
     @Test
