+++ src/test/java/com/woowa/rentcar/K5Test.java
@@ -1,9 +1,10 @@
 package com.woowa.rentcar;
 
+import com.woowa.rentcar.domain.K5;
+
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 class K5Test {
     @Test
