+++ src/test/java/com/woowa/rentcar/SonataTest.java
@@ -1,9 +1,10 @@
 package com.woowa.rentcar;
 
+import com.woowa.rentcar.domain.Sonata;
+
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 class SonataTest {
     @Test
